sld-remote-state/requirements.txt
boto3==1.17.14
botocore==1.20.14
click==7.1.2
fastapi==0.65.2
h11==0.11.0
jmespath==0.10.0
pydantic==1.7.4
pymongo==3.12.1
python-dateutil==2.8.1
s3transfer==0.3.4
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.14.2
urllib3==1.26.5
uvicorn==0.12.2
google-cloud-storage==1.43.0
